Create AccuracyTest.java
AccuracyTest is a small utility used to analyze OCR algorithm accuracy.  

Called from the command line, it scans the user's working directory for any image files, and for a file containing a list of the correct (expected) string outputs from the OCR algorithm.  It grades the algorithm by the number of correct characters matched.